server/rag-python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
faiss-cpu>=1.7.0
transformers>=4.35.0
torch>=2.0.0
pydantic>=2.0.0
python-dotenv==1.0.0
numpy>=1.24.0
requests==2.31.0
huggingface-hub>=0.17.0
